Update patch set 4
Patch Set 4:

I would expect (1) to be the API as a user. That's how a bytes.Buffer behaves, after all, and we're supposed to be a subset of a bytes.Buffer.

Whether or not the "string" type shares the memory or not with the growing []byte is an implementation detail. I suspect that's what Ian was more uncomfortable about.

Ian?

If we don't want to share the growing []byte memory, we can set a "called string" bool flag and then copy the []byte into a freshly-allocated []byte upon the next Write* call.

Patch-set: 4
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>