Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> (1 comment)

I tried it both ways, and leaving that propagation in I saw a fair number of names being copied that were not constants, and I could notice a difference in the debugging quality (in places).  I don't have a test for this difference yet, partly because it was tricky to figure out the cause (I saw it in some parts of the library, not in my test) and partly because I'd have to redo all my stacked CLs with all the changes propagating through that single test.

I also decided to leave it in because we also don't have clear anti-examples yet; perhaps the copy propagation to names is good, except for constants.

Patch-set: 10
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>