Update patch set 1
Patch Set 1: Code-Review-2

Also the lack of the trailing newline while running the actual build steps means that if an error does occur and is printed (by a subprocess), it will be printed after the ... instead of on a new line by itself. That completely sinks this as far as I'm concerned. The subprocesses cannot be run with a half line of output printed, and I don't want to double the number of lines printed just to get times that most people don't care about.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2