Update patch set 1
Patch Set 1:

I just wanted to get something out there to have further discussions. I didn't quite expect the changes to be so pervasive.

The port checks we got for free but can be removed since that wasn't part of the original proposal.
I was also wondering if we are concerned with the amount of duplicate processing that occurs on every no_proxy check.

Patch-set: 1
Reviewer: Gerrit User 5889 <5889@62eb7196-b449-3ce5-99f1-c037f21e1705>