Update patch set 2
Patch Set 2:

Seems fine, but this is more helpful than debug/dwarf generally tries to be. You'd have to do basically the same set of steps to read the type of a variable, which is surely at least as common an operation, but it doesn't have an all-in-one helper to read the attribute and turn it into a Type. I personally think that a more helpful API would be good, but that should be an explicit decision.

Patch-set: 2
Reviewer: Gerrit User 17090 <17090@62eb7196-b449-3ce5-99f1-c037f21e1705>