doc: move plain git introduction in contribution guide
The explanation that plain git can be used is buried within the
installation instructions for git-coverreview. These notes should be
something that proceed installation of the custom tool and be part of
the discussion where the git-codereview tool is first introduced.

The paragraph separator added after the statement about code review in
general and the sentence about the custom tool improves the structure of
the section in general now that it contains more content.

Change-Id: I87ea4e2fa5e31ac16190684ccc8622ae6fa377ca