runtime: new Mode field in cgo symbolizer function arg
When a custom cgo symbolizer function is supplied using
runtime.SetCgoTraceback, the function is called when non-Go
pc values need be symbolized.

Symbolization of non-Go pc values can occur in different
situations including

1) traceback during crash
2) non-crash traceback (e.g. /debug/pprof/goroutines?debug=2)
3) user-requested symbolization through runtime.CallersFrames
   (e.g. symbolization during pprof output formatting)

The case 1 was well recognized and the runtime.SetCgoTraceback
explicitly states that the cgo symbolization function must
be cautious about memory allocation. Crashing during
symbolization during crash is not ideal, so the function must
be safe.

The case 3 adds a new requirement to the cgo symbolization
function. Note runtime.FuncForPC, the predecessor of
runtime.CallersFrames, didn't try to look up cgo symbols,
so this is rather new. The function must be reasonably fast.

The case 2 shares the call path of the case 1 implementation-wise
(traceback.go) but the performance requirement on the cgo
symbolizer function used for this is more similar to the case 3's.
The faster lookup is higher priority.

In some cases, meeting the two requirements (avoiding memory
allocation vs fast) may be challenging. This change introduces
an additional field to the cgoSymbolizer function call argument
and uses it to provide hints on the runtime status.

More flexible, alternative approach could be to allow users
to provide the requirements so the cgoSymbolizer function switches
between different implementations more intelligently based on
the user requirement specification. But, I can't think of an
easy way to allow that without changes in exported runtime APIs.
So, now we start with the two modes: whether runtime is crashing
or not.

The new 'mode' field is added as the last field, so users of
cgo symbolizer that don't care about the different mode of
operations wouldn't be affected.

This is for the alternative approach to #22294
(optimize the cgo symbolizer function).

Change-Id: Ie9cc4f292746d6716bac19897f88f9306fb21fc1