Update patch set 2
Patch Set 2: Code-Review-1

When we discussed this on the issue #22484 I thought the side by side layout as proposed in https://github.com/golang/go/issues/22484#issuecomment-340656701 made it very clear that the left side was the git-codereview way and the right side was the explanation of what the left was doing. This arranges them linearly with "or" separating them. Presenting them in this way introduces choice to the instructions and I believe it will confuse more than anything. 

I'm supportive of adding the alternative instructions but only if it can be done in a way that doesn't make the experience any more complicated for people who are just using the codereview tool (and don't have deep git experience).

Patch-set: 2
Reviewer: Gerrit User 14840 <14840@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1