Update patch set 1
Patch Set 1:

Thanks, but unfortunately this would be a breaking change.  For example, somebody might write cmd.Stdout = os.Stdout.  Then they run the command and wait for it.  With your change, os.Stdout would then be closed.  But the program might still want to write to os.Stdout.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>