runtime/trace: user annotation API
This CL presents the proposed user annotation API skeleton.

Design doc https://goo.gl/iqJfJ3

Implementation CLs are followed.

The API introduces two basic building blocks, "Task" and "Span" for
modeling user-defined logical operation processed by one or more
goroutines.

A task is a logical operation defined by users. An RPC request,
a HTTP request, a file write, a batch job, or any kind of operations
that users may want to create a separate context.Context object.
A task may consists of zero or more subtasks which create a hierarchy.
Processing a task may require multiple goroutines working together.
A Task can start and end in different goroutines.

A span is a time interval in a goroutine. It can be thought as
a Log with a time interval. Spans can be nested. For example,
a function call that calls sub functions can be represented as
a span with subspans.

Users can create a span to record the time interval when a goroutine
is working on a task, by passing the context that represents the task.

For example, it is possible that in the middle of processing a task
for a request (span A), the goroutine may switch briefly to perform a
background task (span B). Span B is nested inside Span A, but span B
and A are unrrelated in terms of tasks (despite the fact that span B
affects span A's latency).

Note we chose to design the API around context.Context so it allows
easier integration with other tracing tools, often designed around
context.Context as well.

In order to create a Task,

   ctx, end := trace.NewContext(ctx, "a task")
   defer end()

Execution tracer tool will measure the time between the task created
and the task ended for the task latency.

To mark that a goroutine is "working" for the task,

   trace.WithSpan(ctx, "name", func(ctx context) {
     /* do something interesting */
   })

Logging adds timestamped log messages to the trace along with
extra info such as the caller goroutine, the category of the log
message, and what task (embedded in the context) is relevant to
this log message.

   trace.Log(ctx, "error", err)

Execution tracer tool can use the key-value messages to classify
tasks and spans. (They are in the followup CLs, not in this CL).

More historical discussion history in golang.org/cl/59572.

Change-Id: I59a937048294dafd23a75cf1723c6db461b193cd