runtime/trace: user annotation API
This CL presents the proposed user annotation API skeleton.

Design doc https://goo.gl/iqJfJ3

Implementation CLs are followed.

The API introduces two basic building blocks, "Task" and "Span" for
modeling user-defined logical operation processed by one or more
goroutines.

Task is a logical operation defined by users. An RPC request,
a HTTP request, a file write, a batch job, or an expensive
function call can be a Task. A task may consists of zero or more
subtasks, which create a hierarchy. Processing a task may require
multiple goroutines to work together. A Task can start and end
in different goroutines.

Span is a time interval that specified when a goroutine works for the
span. Spans can be nested but there is no hierarchy among nested spans.
For hierarchy, create a subtask.

For example, it is possible that in the middle of processing a task
for a request (span A), the goroutine may switch briefly to perform a
background task (span B). Span B is nested inside Span A, but span B
and A are unrrelated in terms of tasks (despite the fact that span B
affects span A's latency).

Note we chose to design the API around context.Context so it allows
easier integration with other tracing tools, often designed around
context.Context as well.

In order to create a Task,

   ctx, end := trace.NewContext(ctx, "a task")
   defer end()

Execution tracer tool will measure the time between the task created
and the task ended for the task latency.

To mark that a goroutine is "working" for the task,

   trace.WithSpan(ctx, "name", func(ctx context) {
     /* do something interesting */
   })

Logging allows to add timestamped key-value type log messages to a Task.

   trace.Log(ctx, "error", err)

Execution tracer tool can use the key-value messages to classify
tasks and spans. (They are in the followup CLs, not in this CL).

More historical discussion history in golang.org/cl/59572.

Change-Id: I59a937048294dafd23a75cf1723c6db461b193cd