Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I'll address some of the issues shortly, but as a main point, I don't want to make this GoLand / any other editor centrinc.
> > > My apologies if this was misinterpreted in any different way, I will try to do better next time.
> > 
> > No problem! I know that wasn't your intent.
> > 
> > I do want to stay clear of this page becoming a laundry list of features.
> > 
> > I suppose there are two kinds of users:
> > * people new to Go that are looking for a suitable editor
> > * people already using one of these editors and want to fully know how they compare
> > 
> > I'd want to tend toward the first, and keeping this page simple and easy-to-digest is key. For the second, do you think we should split into some other section, like "Advanced editing/navigation features"? With more of an explanation of each? Some of the proposed (and current!) descriptions are not easily digestible by people new to Go.
> 
> I agree with Chris.  I'm not sure we even need the matrix - for the reason  mentioned about confusing new users, and for risk that this page starts to look like a recommendation to use a particular editor.
> 
> The same core set of features (format on save, run tests, show coverage, go to definition, etc.) exists in all of these options.  

If we agree all editors have the same set of core features, then we should just not have this table at all.

But if we are to have this table, then I don't see the problem in listing those features (as long as they are relevant to Go users).

> Wouldn't this page be a more valuable resource to new users if it just covered why one might like to use an editor with dedicated Go support, laid out a few options, and pointed users to places where they can get more info?

If you and the others think that's something we should do, then maybe we should have that instead. We can maybe add a small randomization JS for the order as well so that the order is not biased in any way and that's it.

I think Chris's idea of having a separate page / hidden by default comparison is just as good, since it would be all in a single place.

But as it stands, it's unfair to not have it updated with a set of features that everyone worked hard in the past months to add in their tools.

Chris, the Go team should have traffic statistics for that page, is it useful at all? Because another option would be that we have this removed and we can host this page in another place (like the Go Wiki for example?).

Patch-set: 1
Reviewer: Gerrit User 6473 <6473@62eb7196-b449-3ce5-99f1-c037f21e1705>