Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> I do want to stay clear of this page becoming a laundry list of features.
> 
> I suppose there are two kinds of users:
> * people new to Go that are looking for a suitable editor
> * people already using one of these editors and want to fully know how they compare
> 
> I'd want to tend toward the first, and keeping this page simple and easy-to-digest is key. For the second, do you think we should split into some other section, like "Advanced editing/navigation features"? With more of an explanation of each? Some of the proposed (and current!) descriptions are not easily digestible by people new to Go.


I also think this approach is not scalable. This table will be never be completed and I think each editor has it's own style of features. For example in Vim there are text objects and motions, that are quite special to Vim, which allows you to move between parts of your source code in terms of objects. For example it allows you to delete the inner body of a function, let you select it, move it around, etc.. But all this is not bound to a shortcut or mapping, it just blends into Vim environment itself. 

So, in this case, does it means that all other editors lack this feature? I quite don't think this is something that can be represented with this table. I appreciate all the work that goes into this, but going forward I'm not sure this is the way we should approach things.

Patch-set: 2
Reviewer: Gerrit User 5296 <5296@62eb7196-b449-3ce5-99f1-c037f21e1705>