Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Mikioh, any comments?
> 
> have no time to track all netlink implementations and docs, but not keen on returning a pair of parsed messages up to any failure and <non-nil error>. i'd rather see returning parsed and integrity guaranteed messages and <nil>. I mean, I'd like to respect for returning error values.
> 
> fwiw, the route (which grabs routing information base inside the kernel on bsd variants) and internal/socket (which grabs socket control messages on unix variants) packages in x/net repo behave as the latter.


AFAIK, as a language idiom, one should assume that a non-nil error return implies that the returned value should not be used. Thus, I think either return works, and it is up to the caller to decide if/how to use the parsed messages (e.g., to aid debugging by showing which messages were parsed correctly). By returning a nil slice, the code is *ensuring* that the return value can not be used.

Patch-set: 6
Reviewer: Gerrit User 16965 <16965@62eb7196-b449-3ce5-99f1-c037f21e1705>