cmd/go/internal/work: write trivial.c when testing gcc flags
CL 61111 incorrectly disabled this in -n mode,
but if you don't write the input file then the gcc invocation
that follows is guaranteed to fail, and then -n mode will
print different flags.

Or, if the code is actually correct regardless of whether trivial.c
exists, then let's invoke gcc on a non-existent file all the time.

Change-Id: I0a050bfc148fe5a9d430a153d7816b2821277f0d