Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review+1
> 
> Thanks for putting this together! Perfect timing for a project I'm working on.
> 
> One minor thing I encountered was how it handles comments. For example, if I'm trying to rewrite `package user` to `package main` and my source is:
> 
> ```
> package user
> 
> func user() {
> 	println("i'm user")
> }
> 
> // hi world
> // hello ya'll
> // cool story bro
> ```
> 
> The result will be:
> 
> ```
> package // hi world
> // hello ya'll
> // cool story bro
> main
> 
> func user() {
> 	println("i'm user")
> }
> ```
> 
> Which to my surprise does work fine, but it's quite unexpected.
> 
> FWIW, I'm using this code to do the transform (file is an *ast.File from the loader):
> 
> ```
> newAST := rewrite.Apply(file, rewrite.ApplyFunc(func(c *rewrite.Cursor) bool {
> 	if _, ok := c.Parent().(*ast.File); ok && c.Name() == "Name" {
> 		c.Replace(ast.NewIdent("main"))
> 	}
> 	return true
> }), nil)
> 
> var buf bytes.Buffer
> if err := format.Node(&buf, token.NewFileSet(), newAST); err != nil {
> 	return err
> }
> got := buf.String()
> println(got)
> ```

Comments are unfortunately not a "minor thing" and beyond the scope of this library. The problem is that comments are not "attached" to nodes directly but placed based on position, so any automatic rewrites that also need to maintain comments have to do the (tricky!) work to update comment positions. It's a known problem and there's no easy solution (there's go/ast.CommentMap which might help but which has its own problems). Various issues already exist about these problems.

The long-term solution is to attach comments to AST nodes.

Patch-set: 6
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>