Update patch set 1
Patch Set 1:

Thanks for sending this.  For new API like this we prefer to see an issue or a discussion on golang-dev.  We don't want to just see a change.

In this particular case there are many many different variants of CSV.  Rather than try to support them all in the standard library, we prefer the encoding/csv package to stick as close as possible to RFC 4180, as described in the package comment.  So any discussion of a new feature to encoding/csv should start from that point, with the understanding that if the RFC doesn't permit it we are unlikely to add it.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>