archive/tar: partially revert sparse file support
This CL removes all externally facing API for sparse file support.
The internal logic for encoding sparse files was simply commented out
because it is still going to be the same when we add the logic back in.

The OS-specific logic for handling sparse files has been entirely deleted.

Updates #22735

Change-Id: I77842acd39a43de63e5c754bfa1c26cc24687b70