Update patch set 7
Patch Set 7:

Not sure where the trybots are today, but I've tested all.bash by hand on linux-amd64 and darwin-amd64 (so gcc and mac-clang) and they're happy with this.

The old code was indeed wrong for -n, since it required err == nil (err should never be nil with a missing file), but the new without-file test seems to be equivalent to the previous with-file test.

Patch-set: 7
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>