Update patch set 1
Patch Set 1: Code-Review+2

I tested this in conjunction with my write barrier change that ran into the R4 clobbering and everything worked.

I agree that this should probably wait for 1.11, since there's no pressing need for it in 1.10 and it is theoretically possible for this to break something (though it would be *quite* a stretch).

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2