Update patch set 1
Patch Set 1:

> Ping Marti

I was still hoping that someone could give pointers as to the best way to do this.

I'm currently thinking of doing it only for error messages containing two types, and entirely in the runtime. But even that would be a non-trivial amount of work, so I'd rather not give that a go and have to go back to a different method.

Patch-set: 1
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>