Update patch set 5
Patch Set 5:

I'm not completely sure about this. It might be a good idea or it might not. I don't think it's critical to get into Go 1.10, so let's leave the decision on this to Go 1.11. I do think we want to add a note similar to the one in the draft release notes.

Thanks.

Patch-set: 5
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>