Update patch set 1
Patch Set 1:

Is there a convention for when SSA Values should have a Pos?

I thought we were trying to move in the direction of ensuring Values always have a Pos, so this seems like moving in the wrong direction to me.

Patch-set: 1
Reviewer: Gerrit User 5440 <5440@62eb7196-b449-3ce5-99f1-c037f21e1705>