Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is there a convention for when SSA Values should have a Pos?
> 
> I thought we were trying to move in the direction of ensuring Values always have a Pos, so this seems like moving in the wrong direction to me.

I think some values should not have an assigned Pos, at least not in the sense of "visible to the debugger".  Giving these constants a position tends to confuse more than it helps -- stepping in a debugger looks awful, and if a crashed program were somehow stopped at one of these instructions, a naive user would think that it was at line 67, with the implication that all the stuff between entry and line 67 had been executed, which is clearly not the case.

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>