Update patch set 7
Patch Set 7:

This is too late for Go 1.10.

I also need to understand what's going on here. The commit message doesn't tell me enough.

The tree opens again in February for Go 1.11; I'll try to read the GitHub issue and understand the problem before then, but if you want to update the commit message here with more of a story, that'd be great.

Patch-set: 7
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>