Update patch set 1
Patch Set 1: Run-TryBot+1 Code-Review+2

(1 comment)

Thanks for finding this and sending a fix. If you can just add a comment above the select for future reference that would be great. We need to get this in ASAP as the first go1.10 beta is about to be cut. Next steps I'm going to run the try-bots and add Bradfiz to the reviewers in case I missed anything. I'll be marking this as a +2, which means, as soon as the comments are addressed (add comment), it can be submitted (which either myself or Brad will do).
Thanks!

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2
Label: Run-TryBot=+1